Create ticket_price
This app calculate ticket price. The rules are on the below for this app:


Make a program in Java that calculates airfare price based on distance and conditions. Get Distance (KM), age and trip type (One Way, Round Trip) information from the user. Take the fare per distance as 0.10 TL / km. First calculate the total price of the flight and then apply the following discounts to the customer according to the conditions;

The values received from the user must be valid (distance and age must be positive numbers and trip type must be 1 or 2). Otherwise, the user should receive a warning saying "You entered incorrect data !".
If the person is under 12 years old, 50% discount is applied on the ticket price.
If the person is between the ages of 12-24, a 10% discount is applied on the ticket price.
If the person is over 65 years old, a 30% discount is applied on the ticket price.
If the person has selected "Trip Type" as round trip, a 20% discount is applied on the ticket price.